+++ wakepy/_darwin.py
@@ -1,14 +1,15 @@
 from subprocess import Popen, PIPE
 
 # See: https://ss64.com/osx/caffeinate.html
-COMMAND = u"caffeinate"
+COMMAND = "caffeinate"
 
 ARGS = [
     "-d",  # Prevent the display from sleeping.
     "-u",  # Create an assertion to declare that user is active. If the dis-
     # play is off, this option turns the display on and prevents the
     # display from going into idle sleep.
-    "-t 2592000",  # Set timeout to 1-month (default is 5 seconds for -u option)
+    "-t 2592000",  # Set timeout to 1-month
+    # (default is 5 seconds for -u option)
 ]
 
 _process = None
@@ -16,8 +17,8 @@ _process = None
 
 def set_keepawake(keep_screen_awake=False):
     """
-    Set the keep-awake. During keep-awake, the CPU is not allowed to go to sleep
-    automatically until the `unset_keepawake` is called.
+    Set the keep-awake. During keep-awake, the CPU is not allowed to go to
+    sleep automatically until the `unset_keepawake` is called.
 
     Parameters
     -----------